Change of incorrect urls on "Getting Involved" page
There was a wrong url to the modding documentation on the "Getting Involved" page on the plutonium.pw website. This would lead to a 404 Error.
The url was: https://plutonium.pw/modding/
Changed to: https://plutonium.pw/docs/modding/
There was a wrong url to the "Contribution" page on the plutonium.pw website. The would lead to yet another 404 Error.
The url was: https://plutonium.pw/how-can-i-contribute/
Changed to: https://plutonium.pw/docs/how-can-i-contribute/