Links On "Getting Involved" Page
Some links on the "Getting Involved" Page are completely wrong. I previously made a pull request (that was accepted 😀) that attempted to fix some of the false redirects, however due to frequent updates in the change of links to wrong pages, I have decided to make another pull request that changes the Modding Documentation and Contributing hyperlinks.
They previously were:
https://plutonium.pw/modding/ --> Which leads to a 404 page when trying to access the Modding section.
https://plutonium.pw/docs/getting-involved/how-can-i-contribute/ --> Which is a weirdly formatted url that leads to another 404 page when trying to access the How Can I Contribute page.
The previous pull request did not end up changing the modding page, for some reason it still redirects to https://plutonium.pw/modding --> which is a page that does not exist.