Add a note about opening port 3074
This may be basic info to people experienced with BO2 so they may have already opened 3074 to set their NAT type to Open, but I first played BO2 through Pluto, so 3074 was closed, and the server appeared to be functioning normally aside from it not appearing in the server list or being able to be connected to manually. Opening 3074 instantly made the server appear in the server list and made it connectable.